Add support for properly pseudo-escaping backticks in C<code>
In markdown, if the code contains any backticks, we need to deal with
it differently.

If it contains a backtick, we need to find the longest contiguous length,
may it be one ` or two `` or three ``` or even more. We then need to
make sure our code blocks open with some number of backticks that is larger
than this.

If the code starts with a backtick, we must pad it with a space, and likewise
if the code ends with a backtick, we must pad it with a space there as well.

Note: these padded spaces do not show up in the final rendered form and is
the correct way to deal with this.