Use Pod::To::HTML =table rendering
See https://github.com/jpve/perl6-pod-to-markdown/issues/6 for the
motivation for this. Unfortunately it doesn't do everything one
would like, because Rakudo's Pod parsing doesn't process the
formatting:

    https://rt.perl.org/Ticket/Display.html?id=114480

Still, it seems that Pod::To::HTML is the right way to do it
generally, this gives us <caption> support, and ensures that
future improvements to table formatting are automatically
applied.