Use App::Mi6 to build README; move docs to doc/
Because raku --doc crashes when trying to format a Pod/To/* module, we
used to have a hack that would print the Markdown content when you ran
the module as MAIN. This is brittle, and doesn't play well with a
release manager like App::Mi6.

Instead, just accept the state of the world and move the docs out of the
module into doc/Pod/To/Markdown.rakudoc.