Add Pod::Defn handling
The handling of the fake Pod::Defn for older compilers is just copied
from https://github.com/Raku/Pod-To-HTML/issues/43. Probably it isn't
needed anymore, nobody is still using that old of a compiler. But it
doesn't hurt.

The :formatting<B> config is not documented explicitly for =defn, and
I've chosen to use it to format only the term, not the definition.

Fixes #24